server load time reduced to ~1 second
Previously at server start-up every single room would be read from disk
to ensure it existed. Any rooms that did not exist on the disk were then
generated and saved to disk.

Initial server start-up times for full worlds (1K*1K*10) were several
minutes, with subsequent load times being shorter, but still multiple
minutes.

With this commit, rooms are only generated and added to disk when they
are visible to a player or NPC for the first time. This is accomplished
by storing terrain data in memory in a 2D vector of chars, and using it
to generate new rooms when first approached.

Initial and subsequent server start-up times for full worlds are now
approximately one second.