append map type to gen folder
What hasn't changed:

When a terrain or mineral map is created, the intermediate generation
files are placed in a folder called gen_[timestamp]. These intermediate
files exist purely for the server admin to understand the steps taken in
the generation process.

What has changed:

For clarity, this commit appends the type of map being generated to the
name of the intermediate folder.