add NPC backend for multiplayer
Part of #177.

The player can crash the server by approaching NPCs, so until this is
fixed, the line of code that activates NPCs is commented out. For the
brave, this can be enabled in Game::Game().

also:
- generating outbound messages using an Update_Messages object has now
been moved to its own function because the same code needs to be used
from two places (one resulting from player actions, and the other
resulting from NPC actions)
- tweaks to worker NPC code structure