fix a crash
Any NPC that uses the get_new_hostile_id function might look beyond the
edge of the world, at which point the server would crash because no
bounds-check is performed.

This is unfortunately not the cause of #200, because no NPC would ever
near the edge of the world at the time of that issue's creation.